refactor and simplify tests
- separate tests by data type
- shift to non pedantic tests
- remove randomness in favor of single entry
